Fix double variable assignment in initramfs for login
> Can you have a look at 525053 ?
>
> What you are proposing seems to be part of an older problem.
>
>

Indeed, the brokenness has been introduced by this patch, it looks like
the two proposed patches have been applied, and they conflict, hence the
problem I observed.

In that case the best is to revert the first proposed patch. It can be
done with the new patch I attached to this mail.