Handle network swap devices (blocks)
Closes: #756843

And we do not need to worry about fsck here. If the block device really
has a file system that needs a check, mount should sense it.

If the user wants to define an fsck policy, she can anyways define it in
/etc/fstab file

Explicit fsck on each network block device is expensive. Imagine a host
which has 500 LUNs mapped from a SAN Array, with 4 paths per LUN....