Clean up sysvinit ordering w.r.t. NFS.
Move mountnfs-bootclean from Requires-Start to Should-Start. This makes
the init script only depend on mountnfs-bootclean if the administrator
hasn't disabled the latter. Also, have a corresponding Should-Stop
dependency on umountnfs, so that shutdown ordering remains consistent
at all times (systemd's generator doesn't evaluate Should-Stop so this
is safe, i.e. will not cause the deadlock on systemd systems for which
$remote_fs was removed in the first place).