Remove obsolete /var/lib/open-iscsi directory.
/var/lib/open-iscsi was the directory in which open-iscsi 1.0 would
store its database. open-iscsi 2.0 stores it in /etc/iscsi, since it's
actually configuration (and not dynamically modified unless an
administrator actually issues commands).

open-iscsi 1.0 was part of Debian unstable for a short time between the
Sarge (3.1) and Etch (4.0) releases. But Etch (4.0) already came with
open-iscsi 2.0, so this directory is a vestigial relic of old
packaging.

Therefore, just remove the directory in postinst if it's empty; on
current systems it's always empty and the source code contains no
reference to it whatsoever.