Rewrite umountiscsi.sh, make it more robust
This is a complete rewrite of umountiscsi.sh. It now does a much better
job of detecting which of the currently mounted filesystems are on
iSCSI and tries to umount them. It also generates a list of sessions
that contain excluded mounts (default: / on all systems and /usr on
sysvinit systems) and stores them in /run, so that the logic that does
the session logout can use that list to not terminate the session.

The script does not handle every corner case when it comes to stacking,
but should cover and the vast majority of use cases. It is also easy to
add any missing use-cases later on.

Long-term goal should be to not need this script but rather have the
init system take care of all dependencies introduced by stacking, but
we are not there yet.

Closes: #775838