Restart iscsid on upgrades, separate init scripts.
Separate out init scripts and restart iscsid on package upgrades. This
is the canonical way of upgrading open-iscsi that is also supported
upstream. The binary socket protocol that iscsid and iscsiadm use to
communicate is subject to changes between versions, thus the previous
behavior of keeping the old iscsid running was incorrect.

Since a restart of iscsid entails a reconnection of all iSCSI sessions,
try to figure out if all sessions are connected before performing the
upgrade and provide the administrator with an out if that is not the
case. Also, after the upgrade, wait until all iSCSI sessions are
reconnected before proceeding, and display a notice if that wasn't
successful.

debconf is now used to display the prompts and notices before upgrades.
Since we use debconf in preinst (for allowing the administrator to
abort an upgrade if some sessions are failed), we have to add it to
Pre-Depends.

Update README.Debian to provide accurate information and reflect the
current upgrade policy.

We don't use db_input here to ask a regular debconf question but rather
to ask show the user a critical notice if something went wrong.
Therefore, we can safely override this warning.

Closes: #501321