boot: wait long enough for multipath to create mappings
When using multipath, the daemon will race for a lock against udev (and
typically lose), and will wait for 1s before retrying that lock. Only
then the multipath mappings will be added.

The previous sleep of 0.1 seconds is too short to catch that
regularily, so increase it to 3 seconds.

On the other hand, if no volume groups should be activated directly
(because LVM is not used - or the user uses lvmetad and has systemd),
drop out early, because then it's unnecessary, because the boot will be
entirely event-based.