Add simple autopkgtests.
Add very simple autopkgtests that install the open-iscsi package and
then try to see if an initiator name was generated and the iscsid
daemon was successfully started.

It also tests that this works properly on sysvinit systems, where init
scripts instead of native systemd services are used. (Make sure there
are no regressions.)

Closes: #710148