Don't check whether iscsi-iname exists, don't use hardcoded path names.
If postinst is run, we can safely assume that the open-iscsi binaries
are properly installed - and even if not, we use set -e in postinst, so
the call to iscsi-iname will fail the script regardless.

Also, don't use hard-coded path for iscsi-iname, just call it directly,
PATH should be set properly anyway.

Git-Dch: Ignore