Fix handling of multiple kept sessions on shutdown
If multiple sessions were to be kept at shutdown, logout-all.sh didn't
properly read in the list of sessions to be kept in a way it would then
be able utilize them. This is now fixed.

Also fix a cosmetic error in umountiscsi.sh that ls printed some
(harmless) errors.

Closes: #809320